# openapi.yaml
openapi: 3.0.3
info:
  title: AI Summarizer API
  version: 1.0.0
  description: |
    Production-ready REST API for text summarization using Claude AI (Anthropic).

    ## Features
    - Text summarization with configurable length, format, and language
    - Automatic retry with exponential backoff
    - Circuit breaker for resilience
    - Rate limiting
    - Idempotency support
    - Comprehensive monitoring

    ## Authentication
    Currently server-to-server only. API key is configured on the server.

    ## Rate Limits
    - Default: 10 requests per minute per IP
    - Headers: `X-RateLimit-Limit`, `X-RateLimit-Remaining`, `X-RateLimit-Reset`

    ## Idempotency
    Include `Idempotency-Key` header (UUID) to safely retry requests.
    Same key within 24 hours returns cached response without re-processing.

  contact:
    email: support@example.com
  license:
    name: ISC
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.example.com
    description: Production server
...
paths:
  /api/summarize:
    post:
      tags:
        - Summarization
      summary: Summarize text
      description: |
        Summarizes the provided text using Claude AI with configurable options.

        **Idempotency**: Include `Idempotency-Key` header to safely retry.
      operationId: summarizeText
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SummarizeRequest'
            examples:
              basic:
                summary: Basic summarization
                value:
                  text: "Artificial intelligence (AI) is intelligence demonstrated by machines, in contrast to the natural intelligence displayed by humans and animals. Leading AI textbooks define the field as the study of intelligent agents: any device that perceives its environment and takes actions that maximize its chance of successfully achieving its goals."
              withOptions:
                summary: With all options
                value:
                  text: "Artificial intelligence (AI) is intelligence demonstrated by machines, in contrast to the natural intelligence displayed by humans and animals. Leading AI textbooks define the field as the study of intelligent agents: any device that perceives its environment and takes actions that maximize its chance of successfully achieving its goals."
                  length: "short"
                  format: "bullets"
                  language: "Spanish"
      responses:
        '200':
          description: Successful summarization
          headers:
            X-Request-ID:
              $ref: '#/components/headers/X-Request-ID'
            X-RateLimit-Limit:
              $ref: '#/components/headers/X-RateLimit-Limit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/X-RateLimit-Reset'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SummarizeResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/summarize/batch:
    post:
      tags:
        - Summarization
      summary: Batch summarize multiple texts
      description: |
        Summarizes multiple texts concurrently with a concurrency limit of 3.
        Handles partial failures gracefully.

        **Limits**:
        - Maximum 10 texts per batch
        - Concurrency limit: 3 (processes 3 at a time)
        - Each text: 50-10,000 characters

        **Partial Failures**: Returns 207 Multi-Status if some succeed and some fail.
      operationId: batchSummarizeText
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchSummarizeRequest'
            examples:
              basic:
                summary: Basic batch request
                value:
                  items:
                    - text: "First text to summarize. This needs to be at least 50 characters long for validation."
                    - text: "Second text to summarize. This also needs to meet the minimum length requirement."
              withOptions:
                summary: With different options per item
                value:
                  items:
                    - text: "First text to summarize. This needs to be at least 50 characters long."
                      length: "short"
                      format: "bullets"
                    - text: "Second text with different options. Also meeting minimum length."
                      length: "long"
                      format: "paragraph"
                      language: "Spanish"
      responses:
        '200':
          description: All items processed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchSummarizeResponse'
        '207':
          description: Partial success (some items failed)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchSummarizeResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '500':
          description: All items failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchSummarizeResponse'
...
components:
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      description: |
        Unique identifier for idempotent requests (UUID recommended).
        Same key within 24 hours returns cached response.
      required: false
      schema:
        type: string
        format: uuid
        example: "550e8400-e29b-41d4-a716-446655440000"
  headers:
    X-Request-ID:
      description: Unique request identifier for tracing
      schema:
        type: string
        format: uuid
    X-RateLimit-Limit:
      description: Maximum requests allowed in the time window
      schema:
        type: integer
    X-RateLimit-Remaining:
      description: Requests remaining in the current window
      schema:
        type: integer
    X-RateLimit-Reset:
      description: Seconds until the rate limit resets
      schema:
        type: integer
  schemas:
    SummarizeRequest:
      type: object
      required:
        - text
      properties:
        text:
          type: string
          description: Text to summarize
          minLength: 50
          maxLength: 10000
          example: "Artificial intelligence (AI) is intelligence demonstrated by machines..."
        length:
          type: string
          enum: [short, medium, long]
          default: medium
          description: |
            Summary length:
            - `short`: 1-2 sentences
            - `medium`: 3-5 sentences
            - `long`: 2-3 paragraphs
        format:
          type: string
          enum: [bullets, paragraph]
          default: paragraph
          description: Output format
        language:
          type: string
          maxLength: 50
          description: Output language (e.g., "English", "Spanish", "French")
          example: "English"
    SummarizeResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            summary:
              type: string
              description: The generated summary
            metadata:
              type: object
              properties:
                model:
                  type: string
                  description: Claude model used
                  example: "claude-sonnet-4-6"
                usage:
                  type: object
                  properties:
                    inputTokens:
                      type: integer
                    outputTokens:
                      type: integer
                    totalTokens:
                      type: integer
                processingTimeMs:
                  type: integer
                  description: Processing time in milliseconds
                stopReason:
                  type: string
                  description: Reason the generation stopped
        requestId:
          type: string
          format: uuid
          description: Request ID for tracing
    BatchSummarizeRequest:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          minItems: 1
          maxItems: 10
          description: Array of texts to summarize (max 10)
          items:
            type: object
            required:
              - text
            properties:
              text:
                type: string
                minLength: 50
                maxLength: 10000
                description: Text to summarize
              length:
                type: string
                enum: [short, medium, long]
                description: Summary length
              format:
                type: string
                enum: [bullets, paragraph]
                description: Output format
              language:
                type: string
                maxLength: 50
                description: Output language
    BatchSummarizeResponse:
      type: object
      properties:
        success:
          type: boolean
          description: True if all items succeeded
        data:
          type: object
          properties:
            results:
              type: array
              description: Successfully processed items
              items:
                type: object
                properties:
                  index:
                    type: integer
                    description: Index of item in original request
                  data:
                    type: object
                    properties:
                      summary:
                        type: string
                      metadata:
                        type: object
                        properties:
                          model:
                            type: string
                          usage:
                            type: object
                            properties:
                              inputTokens:
                                type: integer
                              outputTokens:
                                type: integer
                              totalTokens:
                                type: integer
                          stopReason:
                            type: string
            errors:
              type: array
              description: Failed items
              items:
                type: object
                properties:
                  index:
                    type: integer
                    description: Index of item in original request
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                      retryable:
                        type: boolean
            summary:
              type: object
              properties:
                total:
                  type: integer
                  description: Total items in batch
                successful:
                  type: integer
                  description: Number of successful items
                failed:
                  type: integer
                  description: Number of failed items
                processingTimeMs:
                  type: integer
                  description: Total processing time
        requestId:
          type: string
          format: uuid
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              description: Machine-readable error code
            message:
              type: string
              description: Human-readable error message
            field:
              type: string
              description: Field that caused the error (if applicable)
        requestId:
          type: string
          format: uuid
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            textTooShort:
              summary: Text too short
              value:
                success: false
                error:
                  code: "TEXT_TOO_SHORT"
                  message: "Text must be at least 50 characters"
                  field: "text"
                  minLength: 50
                  actualLength: 20
                requestId: "550e8400-e29b-41d4-a716-446655440000"
    PayloadTooLarge:
      description: Request payload too large
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    RateLimitExceeded:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Reset:
          $ref: '#/components/headers/X-RateLimit-Reset'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'